src/shared/testing/requirements.txt
# Testing framework dependencies
PyYAML>=6.0
